Add semantic microdata properties
Added semantic microdata to the component mark-up using
`polymer-microdata`:

https://github.com/rgladwell/polymer-microdata

This mixin allows component properties to be set using microdata
markup.

This was a major refactoring of the way the web component works,
including an overhaul of its data schema to make it easier for machines
to consume and reason about monster stats. This work is ongoing and the
schema is still too course. For example, attack damage is unstructured
string that puts average and damage roll as one value, i.e. '13 (2d8 +
4)'.

This closes this ticket:

https://github.com/rgladwell/dnd-stat-block/issues/6

This change also upgraded the component for Polymer 2 and added a
complete set of tests:

https://github.com/rgladwell/dnd-stat-block/issues/3