Support custom stat-blocks
Add support for custom stat-block layouts, such as for location stat
blocks[1] that do not use the strict schema of monster stat-blocks:

https://github.com/rgladwell/vellum-stat-block/issues/8

Monsters are now displayed using the `vellum-monster` custom element,
while the `vellum-stat-block` element provides the basic skeleton for
creating custom stat-blocks.

This change includes:

 * documentation
 * support for the 'official' theme
 * tests

Note: `vellum-monster` inherits from the `vellum-stat-block`. This is
not ideal for a number of reasons, mainly that changes to
`vellum-stat-block` could break `vellum-monster`. Composition, where
`vellum-monster` monster simply calls `vellum-stat-block` would be
preferential but this appears to break the styles and doesn't work in
Chrome.

[1] https://imgur.com/a/aIVfv