Separate attacks into new element
Separate stat block attacks into new custom element to make it easier
to write custom stat blocks mixed with monster stats.

Previously attacks were part of the `<vellum-action>` element, but this
duplicated code in `<vellum-stat>`. Now action-specific code is in the
new `<velum-action>` and `<vellum-stat>` is used for traits and
non-attack actions.

This change also fixes:

https://github.com/grislyeye/vellum-monster/issues/20