Fix: Bars should be wider than stat block body
The top and bottom bars should be slightly wider than the stat block
body:

https://github.com/grislyeye/vellum-monster/issues/15

As part of this change a new CSS variable had to be added to propagate
the stat-block background to the correct shadow DOM element. Also the
box shadow had to be updated to only appear left and right of the stat
block.