Automatically calculate challenge rating
Added the ability to automatically calculate challenge ratings if no
default provided, including tests and documentation. Challenge rating
(and XP) can still be specified and override the calculation. For
example, when playtest indicates is more or less of a challenge than
its CR indicates.

Logs monster challenge rating calculations to table (including
defensive CR, etc.). We attempt to only calculate this if all valid
stats are provided to prevent spamming the console, with limited
success.

Includes support for multiattacks, and AC, hit point, etc. adjustments.

As part of this change we had to modify the 'damage' property of
attacks so that it just provides the die roll. We now calculate the
average damage automatically.

We also had to modify the 'armor' property by splitting it into an 'ac'
property and an optional 'armor' property.

This closes:

https://github.com/grislyeye/vellum-monster/issues/31