Log effective CR calculation results once
In the previous version, effective CR calculations can be logged more
than once due to the web component lifecycle. This change fixes this by
moving the log out of the library and into the web component itself.

As part of this change we had to re-factor some of <vellum-monster>
utility methods, and change the contract for the calculateCr()
function.