Fix: sub heading spacing in html2canvas
Due to a possible issue in html2canvas, there is unwanted spacing
suffixing the first character in the sub-heading when exporting to
image.

This appears to be due to lack of support for the
text-transform: capitalize CSS directive. This component now
explicitly capitalises the the first letter of the first word in the
sub-heading.